+++ src/main/kotlin/com/zepben/evolve/cim/iec61970/base/wires/Breaker.kt
@@ -18,4 +18,10 @@ class Breaker @JvmOverloads constructor(mRID: String = "") : ProtectedSwitch(mRI
      * @return convenience function for detecting if this breaker is part of a substation.
      */
     val isSubstationBreaker get() = substations.isNotEmpty()
+
+    /**
+     * @return convenience function for detecting if this breaker is at the head of a feeder.
+     */
+    val isFeederHeadBreaker get() = normalFeeders.mapNotNull { it.normalHeadTerminal }.any { terminals.contains(it) }
+
 }