+++ src/main/kotlin/com/zepben/evolve/database/sqlite/upgrade/UpgradeRunner.kt
@@ -44,6 +44,7 @@ class UpgradeRunner constructor(
             changeSet30(),
             changeSet31(),
             changeSet32(),
+            changeSet33(),
         ).asUnmodifiable()
 ) {