+++ src/test/kotlin/com/zepben/evolve/cim/iec61970/base/wires/EnergyConsumerTest.kt
@@ -7,7 +7,9 @@
  */
 package com.zepben.evolve.cim.iec61970.base.wires
 
+import com.zepben.evolve.services.common.extensions.typeNameAndMRID
 import com.zepben.evolve.utils.PrivateCollectionValidator
+import com.zepben.testutils.exception.ExpectException
 import com.zepben.testutils.junit.SystemLogExtension
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.*
@@ -53,6 +55,26 @@ internal class EnergyConsumerTest {
         assertThat(energyConsumer.phaseConnection, equalTo(PhaseShuntConnectionKind.Yn))
         assertThat(energyConsumer.q, equalTo(6.7))
         assertThat(energyConsumer.qFixed, equalTo(8.9))
+    }
+
+    @Test
+    internal fun assignsEnergyConsumerToEnergyConsumerPhaseIfMissing() {
+        val energyConsumer = EnergyConsumer()
+        val phase = EnergyConsumerPhase()
+
+        energyConsumer.addPhase(phase)
+        assertThat(phase.energyConsumer, equalTo(energyConsumer))
+    }
+
+    @Test
+    internal fun rejectsEnergyConsumerPhaseWithWrongEnergyConsumer() {
+        val energyConsumer1 = EnergyConsumer()
+        val energyConsumer2 = EnergyConsumer()
+        val phase = EnergyConsumerPhase().apply { energyConsumer = energyConsumer2 }
+
+        ExpectException.expect { energyConsumer1.addPhase(phase) }
+            .toThrow(IllegalArgumentException::class.java)
+            .withMessage("${phase.typeNameAndMRID()} `energyConsumer` property references ${energyConsumer2.typeNameAndMRID()}, expected ${energyConsumer1.typeNameAndMRID()}.")
     }
 
     @Test