+++ src/test/kotlin/com/zepben/evolve/services/diagram/testdata/FillFields.kt
@@ -20,12 +20,8 @@ fun Diagram.fillFields(service: DiagramService, includeRuntime: Boolean = true):
     diagramStyle = DiagramStyle.GEOGRAPHIC
     orientationKind = OrientationKind.NEGATIVE
 
-    for (i in 0..1) {
-        addDiagramObject(DiagramObject().also {
-            it.diagram = this
-            service.add(it)
-        })
-    }
+    for (i in 0..1)
+        addDiagramObject(DiagramObject().also { service.add(it) })
 
     return this
 }
@@ -37,7 +33,7 @@ fun DiagramObject.fillFields(service: DiagramService, includeRuntime: Boolean = 
     diagram?.addDiagramObject(this)
 
     identifiedObjectMRID = "io_mrid"
-    style = DiagramObjectStyle.CONDUCTOR_UNKNOWN
+    style = "CONDUCTOR_UNKNOWN"
     rotation = 1.1
 
     for (i in 0..1)